ZA: Move the attendance class in views.py to its own file
